Add IMDb link monitoring and content fetching feature
A new functionality was added to monitor the `messageCreate` event for IMDb links. If found, the content is fetched through an external API call using `movier`. The fetched content is then embedded into a message that gets replied to the original message. This will allow users to share IMDb movie/series links and automatically display important details about that movie/series.