Added movie type mapping for embed author name
In the 'messageCreate' event, a map object is added to convert the content 'mainType' string into a more readable format. This change applies to the author name section of the embed created when fetching movie details by IMDB ID. This makes the discord embed more user-friendly by showing a more understandable type for the movie.