🔥 (project): economy; read desc...
This took me around an hour, but it was needed.
The economy file got to just under 2000 lines and I was spending more and more time trying to find a specific part I needed to edit.
So I split it into files.
It's probably not 100% functional as-is, but I AM TIRED.

bye